refactor: simplify download status and add DatabaseOperator
- Remove downloading state from DownloadStatus (only persisted states)
- Remove downloadProgress from KomgaBook model
- Add DownloadProgressTracker for in-memory progress display
- Add DatabaseOperator @ModelActor for background SwiftData operations
- Refactor SyncService to delegate DB operations to DatabaseOperator
- Simplify OfflineManager to actor with progress tracking
- Remove unnecessary progress UI from BookDetailView
